Constructor Class
Constructors are used to initialize the state of an object when it is created. They don't return anything explicitly because their primary purpose is to set up the initial state of the object. Instead of returning a value, constructors typically initialize the object's properties or perform any setup tasks necessary for the object to function correctly.